+++ packages/vexide-startup/examples/booting.rs
@@ -4,28 +4,31 @@
 #![no_std]
 
 extern crate alloc;
+
 use alloc::boxed::Box;
 
-use vex_sdk::vexTasksRun;
 use vexide_core::println;
-use vexide_startup::{
-    banner::themes::THEME_DEFAULT, CodeSignature, ProgramFlags, ProgramOwner, ProgramType,
-};
+use vexide_startup::{CodeSignature, ProgramFlags, ProgramOwner, ProgramType};
 
-// SAFETY: This symbol is unique and is being used to start the vexide runtime.
+// SAFETY: This function is unique and is being used to start the vexide runtime.
+// It will be called by the _boot assembly routine after the stack has been setup.
 #[unsafe(no_mangle)]
 unsafe extern "C" fn _start() -> ! {
+    // Setup the heap, zero bss, apply patches, etc...
     unsafe {
-        vexide_startup::startup::<true>(THEME_DEFAULT);
-
-        // Write something to the screen to test if the program is running
-        let test_box = Box::new(100);
-        vex_sdk::vexDisplayRectFill(0, 0, *test_box, 200);
-        println!("Hello, world!");
-        vexTasksRun(); // Flush serial
+        vexide_startup::startup();
     }
 
-    // Exit once we're done.
+    let test_box = Box::new(100); // On the heap to demonstrate allocation.
+    unsafe {
+        // Draw something to the screen to test if the program is running.
+        vex_sdk::vexDisplayRectFill(0, 0, *test_box, 200);
+    }
+
+    // Print some data to the terminal.
+    println!("Hello, world!");
+
+    // Exit the program once we're done.
     vexide_core::program::exit();
 }