+++ packages/vexide-startup/src/patcher/mod.rs
@@ -12,6 +12,17 @@ mod varint_encoding;
 core::arch::global_asm!(include_str!("./overwriter_aeabi_memcpy.S"));
 core::arch::global_asm!(include_str!("./overwriter.S"));
 
+// Linkerscript Symbols
+//
+// All of these external symbols are defined in our linkerscript (link/v5.ld) and don't have
+// real types or values, but a pointer to them points to the address of their location defined
+// in the linkerscript.
+unsafe extern "C" {
+    static mut __patcher_patch_start: u32;
+    static mut __patcher_base_start: u32;
+    static mut __patcher_new_start: u32;
+}
+
 /// Internal patcher state representing what the patcher is attempting to do.
 #[derive(Debug)]
 enum PatcherState {
@@ -97,113 +108,116 @@ pub(crate) unsafe fn patch() {
     // Next four bytes of the patch have to match this version identifier for the patch to be applied.
     const PATCH_VERSION: u32 = 0x1000;
 
-    // Address of the 2mb region containing our reference ("base") binary that the patch will be
-    // built and applied over.
-    //
-    // This is copied to this address by vexide's `_boot` routine that can be found in `lib.rs`.
-    const BASE_START: u32 = 0x07C0_0000;
-    // Address of where we will builder our patched binary at and store it before we overwrite the current
-    // binary with it in stage 2.
-    const NEW_START: u32 = 0x07E0_0000;
-
     /// Load address of patch files.
-    const PATCH_MEMORY: *mut u32 = 0x07A0_0000 as _;
+    const PATCH: *mut u32 = &raw mut __patcher_patch_start;
 
     unsafe {
         // First few bytes contain some important metadata we'll need to setup the patch.
-        let patch_magic = PATCH_MEMORY.read(); // Should be 0xB1DF if the patch needs to be applied.
-        let patch_version = PATCH_MEMORY.offset(1).read(); // Should be 0x1000
-        let patch_len = PATCH_MEMORY.offset(2).read(); // length of the patch buffer
-        let base_binary_len = PATCH_MEMORY.offset(3).read(); // length of the currently running binary
-        let new_binary_len = PATCH_MEMORY.offset(4).read(); // length of the new binary after the patch
+        let patch_magic = PATCH.read(); // Should be 0xB1DF if the patch needs to be applied.
+        let patch_version = PATCH.offset(1).read(); // Shoud be 0x1000
+        let patch_len = PATCH.offset(2).read(); // length of the patch buffer
+        let base_binary_len = PATCH.offset(3).read(); // length of the currently running binary
+        let new_binary_len = PATCH.offset(4).read(); // length of the new binary after the patch
 
-        // Do not proceed with the patch if:
-        // - We have an unexpected PATCH_MAGIC (this is edited after the fact to 0xB2Df intentionally break out of here).
+        // Do not proceed with  patch if:
+        // - We have an unexpected PATCH_MAGIC (We later change this magic to 0xB2DF to intentionally
+        //   trigger this check in order to break the patcher ouf of an infinite loop).
         // - Our patch format version does not match the version in the patch.
         // - There isn't anything to patch.
         if patch_magic != PATCH_MAGIC || patch_version != PATCH_VERSION || patch_len == 0 {
-            // TODO(tropix126): We could reclaim the patch as heap space maybe? Not a high priority.
             return;
         }
 
-        // Overwrite patch magic so we don't re-apply the patch next time.
-        PATCH_MEMORY.write(0xB2DF);
+        // Change patch magic to something invalid so we don't re-apply the patch next time.
+        PATCH.write(0xB2DF);
 
-        // Slice representing our patch contents.
-        let mut patch = core::slice::from_raw_parts(
-            PATCH_MEMORY.offset(5).cast(),
-            patch_len as usize - (size_of::<u32>() * 5),
-        );
-
-        // Slice of the executable portion of the currently running program (this one currently running this code).
-        let mut base = Cursor::new(core::slice::from_raw_parts(
-            BASE_START as *const u8,
+        // Slice of the copy of user program memory we made in vexide's `_boot` routine before any
+        // Rust code had the chance to modify `.bss` or `.data`. This is our base binary.
+        let base = Cursor::new(core::slice::from_raw_parts(
+            (&raw mut __patcher_base_start).cast(),
             base_binary_len as usize,
         ));
 
-        // We'll build our new binary using our patch and base at address 0x07E00000 (NEW_START).
-        let mut new: &mut [u8] =
-            core::slice::from_raw_parts_mut(NEW_START as *mut u8, new_binary_len as usize);
+        // Slice of our patch contents. We offset by 20 bytes to skip the metadata inserted by cargo-v5 and bidiff.
+        let patch = core::slice::from_raw_parts(
+            PATCH.offset(5).cast(),
+            patch_len as usize - (size_of::<u32>() * 5),
+        );
 
-        // Build the new binary, using `base` and `patch` as a reference.
-        //
-        // This is essentially a port of <https://github.com/divvun/bidiff/blob/main/crates/bipatch/src/lib.rs>
+        // This is a 2mb slice of uninitialized memory that we've reserved for building the new binary in.
+        let new = core::slice::from_raw_parts_mut(
+            (&raw mut __patcher_new_start).cast(),
+            new_binary_len as usize,
+        );
 
-        let mut buf = [0u8; 4096];
-
-        let mut state = PatcherState::Initial;
-
-        while !new.is_empty() {
-            let processed = match state {
-                PatcherState::Initial => {
-                    state = PatcherState::Add(patch.read_varint().unwrap());
-                    0
-                }
-                PatcherState::Add(add_len) => {
-                    let n = add_len.min(new.len()).min(buf.len());
-
-                    let out = &mut new[..n];
-                    base.read_exact(out).unwrap();
-
-                    let dif = &mut buf[..n];
-                    patch.read_exact(dif).unwrap();
-
-                    for i in 0..n {
-                        out[i] = out[i].wrapping_add(dif[i]);
-                    }
-
-                    state = if add_len == n {
-                        let copy_len: usize = patch.read_varint().unwrap();
-                        PatcherState::Copy(copy_len)
-                    } else {
-                        PatcherState::Add(add_len - n)
-                    };
-
-                    n
-                }
-                PatcherState::Copy(copy_len) => {
-                    let n = copy_len.min(new.len());
-
-                    let out = &mut new[..n];
-                    patch.read_exact(out).unwrap();
-
-                    state = if copy_len == n {
-                        let seek: i64 = patch.read_varint().unwrap();
-                        base.seek(SeekFrom::Current(seek)).unwrap();
-
-                        PatcherState::Initial
-                    } else {
-                        PatcherState::Copy(copy_len - n)
-                    };
-
-                    n
-                }
-            };
-
-            new = &mut new[processed..];
-        }
+        // Build the new binary using `base` and `patch` as a reference.
+        bipatch(base, patch, new);
 
         // Jump to the stage 2 overwriter routine to handle the rest.
         core::arch::asm!("b __patcher_overwrite", options(noreturn));
+    }
+}
+
+/// `bipatch` Algorithm
+///
+/// This function writes into `new` given an older `base` reference buffer and a [bidiff]-compatible
+/// differential patch buffer that is designed to be applied over the base to build a new binary.
+///
+/// [bidiff]: https://github.com/divvun/bidiff
+///
+/// This is essentially a port of <https://github.com/divvun/bidiff/blob/main/crates/bipatch/src/lib.rs>
+// NOTE: LLVM should always inline this function since it's only called once.
+fn bipatch<B: Read + Seek, P: Read>(mut old: B, mut patch: P, mut new: &mut [u8]) {
+    let mut buf = [0u8; 4096];
+    let mut state = PatcherState::Initial;
+
+    while !new.is_empty() {
+        let processed = match state {
+            PatcherState::Initial => {
+                state = PatcherState::Add(patch.read_varint().unwrap());
+                0
+            }
+            PatcherState::Add(add_len) => {
+                let n = add_len.min(new.len()).min(buf.len());
+
+                let out = &mut new[..n];
+                old.read_exact(out).unwrap();
+
+                let dif = &mut buf[..n];
+                patch.read_exact(dif).unwrap();
+
+                for i in 0..n {
+                    out[i] = out[i].wrapping_add(dif[i]);
+                }
+
+                state = if add_len == n {
+                    let copy_len: usize = patch.read_varint().unwrap();
+                    PatcherState::Copy(copy_len)
+                } else {
+                    PatcherState::Add(add_len - n)
+                };
+
+                n
+            }
+            PatcherState::Copy(copy_len) => {
+                let n = copy_len.min(new.len());
+
+                let out = &mut new[..n];
+                patch.read_exact(out).unwrap();
+
+                state = if copy_len == n {
+                    let seek: i64 = patch.read_varint().unwrap();
+                    old.seek(SeekFrom::Current(seek)).unwrap();
+
+                    PatcherState::Initial
+                } else {
+                    PatcherState::Copy(copy_len - n)
+                };
+
+                n
+            }
+        };
+
+        new = &mut new[processed..];
     }
 }